\name{find_cit}
\alias{find_cit}

\title{
find_cit
}
\description{
Search function for citations (text between parenthesis). 
}
\usage{
find_cit(file_in, file_out)
}

\arguments{
  \item{file_in}{
Filename input, in plain text.
}
  \item{file_out}{
Filename output.
}
}

\value{
File with citations.
}

\author{
Marco Fracassetti
}

\examples{
# file_in="/home/marco/Dropbox/art/maydica/cit"
# file_out="/home/marco/Dropbox/art/maydica/cit_ris"
# find_cit(file_in,file_out)
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
