\name{revcomp}
\alias{revcomp}

\title{
revcomp
}
\description{
Reverse and complement of a sequence.
}
\usage{
revcomp(seq, lowercase = FALSE)
}

\arguments{
  \item{seq}{
A string of DNA.
}
  \item{lowercase}{
If TRUE output in lowercase, (default FALSE).
}
}

\value{
A string with reverse and complement DNA.
}

\author{
Marco Fracassetti
}

\examples{
#primer_R=revcomp("CTCTACATCATCGCGCAGTG")
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
