\name{trascritti}
\alias{trascritti}

\title{
trascritti
}
\description{
The function creates a html page with colored transcripts.
}
\usage{
trascritti(file_gen, file_html)
}

\arguments{
  \item{file_gen}{
Filename input, downloaded from maizesequence in genbank format.
}
  \item{file_html}{
Filename output.
}
}

\value{
A html page with each transcript with a different color.
}
\references{
http://www.maizesequence.org/
}
\author{
Marco Fracassetti
}

\examples{
# file_gen="/home/marco/RT_PCR/ensembl.txt"
# file_html="/home/marco/RT_PCR/prova.html"
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
